[#9] [REFACTOR] 도메인 구조 변경 및 엔티티 매핑 수정
- 기존의 document와 approval을 합쳐 도메인 구조를 변경하였습니다.
- 도메인 이름과 contoller의 이름을 일치시켰습니다.
- 엔티티 매핑에서의 연관 관계를 해제하였습니다.
- 엔티티 생성자에 대한 접근 제어자를 변경했습니다.
- 문서의 state필드에 대한 enum을 생성했습니다.
- Builder패턴에서 private가 아닌 부분들을 수정하였습니다.
- 복합키에서의 @Data 어노테이션을 수정하였습니다.
