--- src/main/java/mashup/backend/hoowook_hoowwok/approval/controller/SignController.java
@@ -1,9 +1,0 @@
-package mashup.backend.hoowook_hoowwok.approval.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequiredArgsConstructor
-@RestController
-public class SignController {
-}
