--- src/main/java/mashup/backend/hoowook_hoowwok/approval/domain/ApprovalRepository.java
@@ -1,7 +1,0 @@
-package mashup.backend.hoowook_hoowwok.approval.domain;
-
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ApprovalRepository extends JpaRepository<Approval, ApprovalId> {
-}
