--- src/main/java/mashup/backend/hoowook_hoowwok/approval/service/SignService.java
@@ -1,9 +1,0 @@
-package mashup.backend.hoowook_hoowwok.approval.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@RequiredArgsConstructor
-@Service
-public class SignService {
-}
