+++ src/main/java/mashup/backend/hoowook_hoowwok/document/repository/ApprovalRepository.java
@@ -1,0 +1,8 @@
+package mashup.backend.hoowook_hoowwok.document.repository;
+
+import mashup.backend.hoowook_hoowwok.document.domain.approval.Approval;
+import mashup.backend.hoowook_hoowwok.document.domain.approval.ApprovalId;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+interface ApprovalRepository extends JpaRepository<Approval, ApprovalId> {
+}
