+++ src/main/java/mashup/backend/hoowook_hoowwok/document/repository/DocumentRepository.java
@@ -1,5 +1,6 @@
-package mashup.backend.hoowook_hoowwok.document.domain;
+package mashup.backend.hoowook_hoowwok.document.repository;
 
+import mashup.backend.hoowook_hoowwok.document.domain.document.Document;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface DocumentRepository extends JpaRepository<Document, Long> {
